requirements.txt
torch>=2.2.1
typer>=0.15.1
fastapi>=0.115.6
uvicorn>=0.34.0
matplotlib>=3.9.2
numpy>=2.2.1
hydra-core>=1.3.2
loguru>=0.7.2
wandb>=0.19.1
pytorch-lightning>=2.5.0